add city if not written
