+++ frontend/src/components/FormRoute/FormRoute.module.css
@@ -5,36 +5,6 @@
   gap: 30px;
   padding: 24px 15px 36px 15px;
 }
-.text {
-  flex-shrink: 0;
-  font-weight: 700;
-  font-size: 18px;
-  line-height: 20px;
-  margin-right: 30px;
-}
-
-.category {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 10px;
-  flex-wrap: wrap;
-}
-
-.select {
-  max-width: 370px;
-  height: 24px;
-  width: 100%;
-  background: #ffffff;
-  border: 1px solid #c4c4c4;
-  border-radius: 50px;
-  text-align: center;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 16px;
-  text-align: center;
-}
 
 .btnSubmit {
   display: inline-block;
@@ -50,19 +20,6 @@
 
 .btnSubmit:hover {
   transform: scale(1.1);
-}
-
-.errorCategory,
-.errorLocation {
-  position: absolute;
-  top: -10px;
-  left: 0;
-  font-size: 10px;
-  color: red;
-}
-.errorLocation {
-  top: 10px;
-  left: 30px;
 }
 
 @media (max-width: 640px) {
