+++ frontend/src/components/InputCategories/InputCategories.module.css
@@ -1,0 +1,45 @@
+.errorCategory {
+  position: absolute;
+  top: -10px;
+  left: 0;
+  font-size: 10px;
+  color: red;
+}
+
+.category {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 10px;
+  flex-wrap: wrap;
+}
+
+.select {
+  max-width: 370px;
+  height: 24px;
+  width: 100%;
+  background: #ffffff;
+  border: 1px solid #c4c4c4;
+  border-radius: 50px;
+  text-align: center;
+  font-weight: 400;
+  font-size: 14px;
+  line-height: 16px;
+  text-align: center;
+}
+
+.text {
+  flex-shrink: 0;
+  font-weight: 700;
+  font-size: 18px;
+  line-height: 20px;
+  margin-right: 30px;
+}
+
+@media (max-width: 640px) {
+  .text {
+    font-size: 16px;
+    line-height: 1.1;
+  }
+}
