+++ frontend/src/components/InputCategories/index.jsx
@@ -1,0 +1,2 @@
+import { InputCategories } from './InputCategories';
+export default InputCategories;
